refactor: :truck: restructure all schema related logic
